Refactor project into package modules
