+++ poseidon/packer.py
@@ -1,0 +1,8 @@
+class Packer:
+    @staticmethod
+    def pack(code: bytes):
+        print("[Packer] Polymorphic packer produced XOR+LZ4 stub")
+
+    @staticmethod
+    def unpack(stub: bytes):
+        print("[Packer] Unpacked stub into memory")
