+++ poseidon/shadow_move.py
@@ -1,0 +1,30 @@
+class ShadowMove:
+    def __init__(self):
+        self.cred_vault = {}
+        self.network_map = self._discover_network()
+
+    def _discover_network(self):
+        return {
+            "subnets": ["192.168.1.0/24", "10.0.0.0/8"],
+            "domain_controllers": ["dc01.corp.local"],
+            "critical_assets": ["sql01.corp.local", "fileserver01"]
+        }
+
+    def harvest_credentials(self):
+        self.cred_vault = {
+            "Administrator": "HASHEDPASSWORD1",
+            "svc_account": "HASHEDPASSWORD2"
+        }
+        print("[SHADOW] Simulated credential harvesting")
+
+    def pass_the_ticket(self):
+        print("[SHADOW] Simulated Pass-the-Ticket attack")
+
+    def kerberoast(self):
+        print("[SHADOW] Simulated Kerberoasting")
+
+    def ldap_enumeration(self):
+        print("[SHADOW] Simulated LDAP/AD enumeration")
+
+    def pth(self):
+        print("[SHADOW] Simulated Pass-the-Hash attack")
